Retoques y arreglos en vistas de usuarios
