+++ Web/client/src/App.js
@@ -11,39 +11,17 @@ import ResultCategory from './components/Progress';
 import UserDetails from './components/UserDetails'
 
 function App() {
-  const [filterValue, setFilterValue] = useState('')  
-  const users = [
-      {uid:1, name:"Benja", puntos:12000,  mail:"Mail123@testing.com"},
-      {uid:2, name:"Martin", puntos:11000, mail:"Mail456@testing.com"},
-      {uid:3, name:"Tomas", puntos:10000, mail:"Mail789@testing.com"},
-      {uid:4, name:"Juan", puntos:9000,    mail:"Mail234@testing.com"},
-      {uid:5, name:"Franco", puntos:8000,  mail:"Mail567@testing.com"},
-      {uid:6, name:"Joel", puntos:7000,    mail:"Mail890@testing.com"},
-      {uid:7, name:"Cassoli", puntos:7000, mail:"Mail012@testing.com"},
-      {uid:8, name:"Nico", puntos:7000,    mail:"Mail654@testing.com"},
-  ]
-  const waitingUsers = [
-      {uid:1, name:"Benja1", puntos:12000,  mail:"Mail123@testing.com"},
-      {uid:2, name:"Martin2", puntos:11000, mail:"Mail456@testing.com"},
-      {uid:3, name:"Tomas3", puntos:10000, mail:"Mail789@testing.com"},
-      {uid:4, name:"Juan4", puntos:9000,    mail:"Mail234@testing.com"},
-      {uid:5, name:"Franco5", puntos:8000,  mail:"Mail567@testing.com"},
-      {uid:6, name:"Joel6", puntos:7000,    mail:"Mail890@testing.com"},
-      {uid:7, name:"Cassoli7", puntos:7000, mail:"Mail012@testing.com"},
-      {uid:8, name:"Nico8", puntos:7000,    mail:"Mail654@testing.com"},
-  ]
   return (
     <Router>
         <Route exact path='/adminUsers' render={() => (
               <div>
-                <Filtrar filterValue={filterValue} setFilterValue={setFilterValue}/>
-                <ListaUsers users={users} filterValue={filterValue}/>
+               {/*  <Filtrar filterValue={filterValue} setFilterValue={setFilterValue}/> */}
+                <ListaUsers/>
               </div>
         )}/>
         <Route exact path='/adminWaitingUsers' render={() => (
               <div>
-                <Filtrar filterValue={filterValue} setFilterValue={setFilterValue}/>
-                <ListaUserWait  waitingUsers={waitingUsers} filterValue={filterValue}/>
+                <ListaUserWait/>
               </div>
         )}/>
        <Route exact path='/login' component={Login} />
